requirements.txt
streamlit
torch
sentence-transformers
Pillow
unidecode
